style: moving classes to their own files
