+++ Assets/Mirage/Runtime/SocketLayer/Connection.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 16fa5bc997f0e4f4f8f60af1fc698f6d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
