+++ Assets/Mirage/Runtime/SocketLayer/Connection/Connection.cs.meta
