+++ Assets/Mirage/Runtime/SocketLayer/Connection/IConnection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8482d64762097eb42aefc979aa42fbe3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
