+++ Assets/Mirage/Runtime/SocketLayer/ConnectionTrackers/ConnectingTracker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2659ac76cf94db4eb9bf5287269ce37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
