+++ Assets/Mirage/Runtime/SocketLayer/ConnectionTrackers/DisconnectedTracker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b967fd482d3b1724bb8147a435d061d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
