+++ Assets/Mirage/Runtime/SocketLayer/ConnectionTrackers/TimeoutTracker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f328e50f9867240489806e84a105dc20
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
